[refactor] 메인 페이지 컴포너트 분리
